backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["isra el <vernu1997@gmail.com>"]
# readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.0"
django-split-settings = "^1.2.0"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
